refactor: rename _change_event_hander.py to _change_event_handler.py
Fix pre-existing typo in filename and update all references.
